Reuse Middleware application DTOs for the Quick Quote with Credit Commitment request/response
